requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.18
supabase==2.11.0
deepface==0.0.93
onnxruntime>=1.17.0
Pillow==11.1.0
numpy>=1.26.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
scikit-learn>=1.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
